refactor(example): Add example domain states and actions, update routes and controllers
