+++ Business/Example/Application/Queries/ListExamplesQuery.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace Business\Example\Application\Queries;
+
+use Illuminate\Contracts\Database\Eloquent\Builder;
+use Illuminate\Http\Request;
+use Business\Example\Domain\Models\Example;
+use Spatie\QueryBuilder\AllowedFilter;
+use Spatie\QueryBuilder\QueryBuilder;
+
+class ListExamplesQuery extends QueryBuilder
+{
+    public function __construct(Request $request)
+    {
+        $query = $this->baseQuery();
+
+        parent::__construct($query, $request);
+
+        $this->allowedIncludes('user');
+
+        $this->allowedFilters(
+            AllowedFilter::exact('id'),
+            AllowedFilter::partial('name'),
+        );
+    }
+
+    private function baseQuery(): Builder
+    {
+        return Example::with('user');
+    }
+}
