+++ Business/Example/Application/Requests/Request.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Business\Example\Application\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class Request extends FormRequest
+{
+    public function authorize()
+    {
+        return true;
+    }
+
+    protected function prepareForValidation()
+    {
+        // do something
+    }
+}
