+++ Business/Example/Domain/States/Completed.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Business\Example\Domain\States;
+
+class Completed extends ExampleState
+{
+    public static string $state = 'completed';
+
+    public static string $name = 'completed';
+}
