+++ Business/Example/Domain/States/Draft.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Business\Example\Domain\States;
+
+class Draft extends ExampleState
+{
+    public static string $state = 'draft';
+
+    public static string $name = 'draft';
+}
