+++ Business/Example/Domain/States/Failed.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Business\Example\Domain\States;
+
+class Failed extends ExampleState
+{
+    public string $state = 'failed';
+
+    public static string $name = 'failed';
+}
