+++ Business/Example/Domain/States/Pending.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Business\Example\Domain\States;
+
+class Pending extends ExampleState
+{
+    public static string $state = 'pending';
+
+    public static string $name = 'pending';
+}
