--- PrototypeOne/Application/Requests/EditExampleRequest.php
@@ -1,30 +1,0 @@
-<?php
-
-namespace PrototypeOne\Application\Requests;
-
-class EditExampleRequest extends Request
-{
-    public function rules()
-    {
-        return [
-            'name' => [
-                'required',
-                'string',
-                'max:255',
-                'unique:examples,name,' . $this->route('example')->id,
-            ],
-            'description' => [
-                'nullable',
-                'string',
-                'max:255',
-            ],
-        ];
-    }
-
-    public function attributes()
-    {
-        return [
-            'name' => trans('messages.prototypes.models.example.name'),
-        ];
-    }
-}
