--- PrototypeOne/Application/Requests/Request.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace PrototypeOne\Application\Requests;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class Request extends FormRequest
-{
-    public function authorize()
-    {
-        return true;
-    }
-
-    protected function prepareForValidation()
-    {
-        // do something
-    }
-}
