--- PrototypeOne/Application/Resources/ExampleResource.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace PrototypeOne\Application\Resources;
-
-use Illuminate\Http\Resources\Json\JsonResource;
-
-/** @mixin \PrototypeOne\Domain\Models\Example */
-class ExampleResource extends JsonResource
-{
-    /**
-     * Transform the resource into an array.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return array
-     */
-    public function toArray($request): array
-    {
-        return  [
-            'id' => $this->id,
-            'name' => $this->name,
-            'description' => $this->description,
-            'state' => $this->state,
-            'created_at' => $this->created_at,
-            'updated_at' => $this->updated_at,
-            'user' => new UserResource($this->user),
-        ];
-    }
-}
