--- PrototypeOne/Domain/Actions/CreateExampleAction.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace PrototypeOne\Domain\Actions;
-
-use PrototypeOne\Domain\DataTransferObjects\ExampleDataTransferObject;
-use PrototypeOne\Domain\Models\Example;
-
-class CreateExampleAction
-{
-    public function execute(ExampleDataTransferObject $exampleDataTransferObject): Example
-    {
-        $example = new Example();
-        $example->fill($exampleDataTransferObject->toArray());
-        $example->save();
-
-        return $example;
-    }
-}
