--- PrototypeOne/Domain/States/Completed.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace PrototypeOne\Domain\States;
-
-class Completed extends ExampleState
-{
-    public static string $state = 'completed';
-
-    public static string $name = 'completed';
-}
