--- PrototypeOne/Domain/States/Draft.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace PrototypeOne\Domain\States;
-
-class Draft extends ExampleState
-{
-    public static string $state = 'draft';
-
-    public static string $name = 'draft';
-}
