--- PrototypeOne/Domain/States/Failed.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace PrototypeOne\Domain\States;
-
-class Failed extends ExampleState
-{
-    public string $state = 'failed';
-
-    public static string $name = 'failed';
-}
