--- PrototypeOne/Domain/States/Pending.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace PrototypeOne\Domain\States;
-
-class Pending extends ExampleState
-{
-    public static string $state = 'pending';
-
-    public static string $name = 'pending';
-}
